Fix validation by using the valid property and fill in required mmsi|uuid|url
Tv4 validation has a bug https://github.com/geraintluff/tv4/issues/128 that causes validation
to fail in some cases. Schema has a workaround for this bug, but to use the workaround
you need to check the valid property and not rely on errors lenght.

Nowadays either uuid, mmsi or url is needed under vessel root.