Move tests to individual files
Comparing to a predefined result makes the tests brittle. In
this case schema changes related to source handling broke all the
parsing tests. Having a single file to test parsers parsing
functionality seems silly.

This change moves asserts to sentence specific tests with a bit more detailed
asserts for the result as well as validation against the schema.