Refactor library to not use Promises internally
The use of promises inside the parser-nmea0183 library forces caller into an
asynchronous mode but we do not actually need Promises here since all the data
can be parsed/calculated immediately.

This change should be completely backwards compatible for callers. The only
difference will be that the event will be emitted immediately when calling the
parse() function, instead of being called during the next tick. Hopefully this
will not impact anyone.

A new `Parser.parseImmediate()` function is introduced to return a SignalK Delta
object from a NMEA sentence.