refactor: do not pass parser to hooks, just the session
This changes the hook API. Instead of receiving a reference to the
parser (which only VDM used for AIS sessions) and the sentence parts as
the second argument, the hooks now receive the sentence parts first and
the session as the second argument.

Because most hooks do not use it, they can just ignore the second
argument making them completely stateless.