fix: HDG deviation handling
Previously deviation was not handled at all. Now if there is
deviation it is applied to create nav.headingMagnetic and in
addition nav.headingCompass is output. This PR also adds
deviation to the output.

If there is no deviation value the output is as previously:
the sensor value in the first field is output as is as
nav.headingMagnetic and no nav.headingCompass. This is
a bit backwards, as what is being output is nav.headingCompass,
but I think changing that would warrant a major version
change for being so incompatible with the current behavior.